+++ Tests/Profiler/Storage/RedisProfilerStorageTest.php
@@ -13,15 +13,24 @@ namespace Symfony\Component\HttpKernel\Tests\Profiler;
 
 use Snc\RedisBundle\Profiler\Storage\RedisProfilerStorage;
 use Snc\RedisBundle\Tests\Profiler\Storage\Mock\RedisMock;
+use Symfony\Component\HttpKernel\Kernel;
 use Symfony\Component\HttpKernel\Profiler\Profile;
 use PHPUnit\Framework\TestCase;
 
+/**
+ * @group legacy
+ */
 class RedisProfilerStorageTest extends TestCase
 {
     protected static $storage;
 
     protected function setUp()
     {
+        if (Kernel::VERSION_ID >= 40400) {
+            $this->markTestSkipped('Redis profiler storage is not supported since Symfony 4.4');
+        }
+
+
         $redisMock = new RedisMock();
         $redisMock->connect('127.0.0.1', 6379);